requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pyee==7.0.4
pylint==2.6.0
pyppeteer==0.2.2
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.1
rq==1.7.0
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
